context-menu launches http-like urls within comments
Records and manages URLs the same ways as misspelled region suggestions, with new public method to access them.

Context menu item gets added in racket.rkt (not necessarily where that code should end up).

I think the context menu launching should be the default, to prevent surprises for users who don't know about the functionality, or it could require specifically setting a new preference that is off by default.

Other than refactoring the duplicated code (which I'm content to do, it's quick and fun when I have a moment), the look-and-feel, along with any preference configuration of that, is very open for anyone who wants to contribute.